Update SPIRV-Tools and fix spacing to match MS.
