+++ include/dxc/Support/FileIOHelper.h
@@ -9,7 +9,6 @@
 //                                                                           //
 ///////////////////////////////////////////////////////////////////////////////
 
-
 #pragma once
 
 #include "Global.h"
