+++ include/llvm/Support/FileSystem.h
@@ -96,6 +96,7 @@ public:
 
   const std::error_code& error_code() const { return ec; }
 };
+
 // HLSL Change Ends
 
 /// An enumeration for the file system's view of the type.
