+++ lib/HLSL/DxilOperations.cpp
@@ -25,6 +25,7 @@ using namespace llvm;
 using std::vector;
 using std::string;
 
+
 namespace hlsl {
 
 using OC = OP::OpCode;
