+++ lib/IR/Function.cpp
@@ -33,7 +33,6 @@
 #include "llvm/Support/RWMutex.h"
 #include "llvm/Support/StringPool.h"
 #include "llvm/Support/Threading.h"
-
 using namespace llvm;
 
 // Explicit instantiations of SymbolTableListTraits since some of the methods
