+++ lib/Support/Regex.cpp
@@ -17,7 +17,6 @@
 #include "llvm/ADT/StringRef.h"
 #include "llvm/ADT/Twine.h"
 #include <string>
-
 using namespace llvm;
 
 Regex::Regex(StringRef regex, unsigned Flags) {
