+++ tools/clang/lib/AST/ExprClassification.cpp
@@ -19,7 +19,6 @@
 #include "clang/AST/ExprCXX.h"
 #include "clang/AST/ExprObjC.h"
 #include "llvm/Support/ErrorHandling.h"
-
 using namespace clang;
 
 typedef Expr::Classification Cl;
